all another selenium tests

+++ src/test/java/com/test/webTest/Tests.java
@@ -3,6 +3,7 @@ package com.test.webTest;
 import com.test.Main;
 import com.test.entity.Conference;
 import com.test.entity.ConferenceRoom;
+import com.test.service.ConferenceParticipantService;
 import com.test.service.ConferenceRoomService;
 import com.test.service.ConferenceService;
 import com.test.webTest.pages.*;
@@ -40,13 +41,16 @@ public class Tests {
     @Autowired
     ConferenceRoomService conferenceRoomService;
 
+    @Autowired
+    ConferenceParticipantService conferenceParticipantService;
+
     private static WebDriver driver;
 
     @BeforeClass
     public static void setup(){
         //set driver
-        System.setProperty("webdriver.chrome.driver","/home/alexandr/IdeaProjects/Test_task_Back-end/src/test/java/com/test/webTest/chromedriver/chromedriver");
-//        System.setProperty("webdriver.chrome.driver","C:\\Users\\IMTOP\\IdeaProjects\\test\\src\\test\\java\\com\\test\\webTest\\chromedriver\\chromedriver.exe");
+//        System.setProperty("webdriver.chrome.driver","/home/alexandr/IdeaProjects/Test_task_Back-end/src/test/java/com/test/webTest/chromedriver/chromedriver");
+        System.setProperty("webdriver.chrome.driver","C:\\Users\\IMTOP\\IdeaProjects\\test\\src\\test\\java\\com\\test\\webTest\\chromedriver\\chromedriver.exe");
 
         //set language
         ChromeOptions options = new ChromeOptions();
@@ -71,37 +75,44 @@ public class Tests {
 
     @Test
     public void addNewConference() {
-        driver.get("http://localhost:8080/conferences");
-        ConferencePage conferencePage = new ConferencePage(driver);
+        try {
+            driver.get("http://localhost:8080/conferences");
+            ConferencePage conferencePage = new ConferencePage(driver);
+            conferencePage.addConference();
 
-        //set data
-        conferencePage.addConference();
-        driver.get("http://localhost:8080/addConference");
-        AddConferencePage addConferencePage = new AddConferencePage(driver);
-        addConferencePage.enterConferenceName("Test Conference", "12-12-2020", "15:00");
-        List actual = addConferencePage.getDataForVerification();
-        addConferencePage.clickSubmit();
+            driver.get("http://localhost:8080/addConference");
+            AddConferencePage addConferencePage = new AddConferencePage(driver);
+            addConferencePage.enterConferenceName("Test Conference", "12-12-2020", "15:00");
+            List actual = addConferencePage.getDataForVerification();
+            addConferencePage.clickSubmit();
 
-        //validate data
-        List expected =  asList("Test Conference", "2020-12-12T15:00");
-        assertEquals(expected, actual);
-
-        //delete test conference
-        conferenceService.deleteByName("Test Conference");
+            List expected = asList("Test Conference", "2020-12-12T15:00");
+            assertEquals(expected, actual);
+        } finally {
+            if (conferenceService.getConferenceByName("Test Conference") != null) {
+                conferenceService.deleteByName("Test Conference");
+            }
+        }
     }
 
     @Test
     public void changeStatus() {
         conferenceService.create(new Conference("Test Conference",
                 LocalDateTime.parse("2020-12-12 13:00", DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm"))));
-        driver.get("http://localhost:8080/conferences");
-        ConferencePage conferencePage = new ConferencePage(driver);
-        conferencePage.setStatusCanceled();
-        String canceledStatus = conferencePage.checkChangedStatus();
-        conferencePage.setStatusActive();
-        String activeStatus = conferencePage.checkChangedStatus();
-        conferenceService.deleteByName("Test Conference");
-        assertEquals(Arrays.asList("canceled", "active"), Arrays.asList(canceledStatus, activeStatus));
+
+        try {
+            driver.get("http://localhost:8080/conferences");
+            ConferencePage conferencePage = new ConferencePage(driver);
+            conferencePage.setStatusCanceled();
+            String canceledStatus = conferencePage.checkChangedStatus();
+            conferencePage.setStatusActive();
+            String activeStatus = conferencePage.checkChangedStatus();
+            assertEquals(Arrays.asList("canceled", "active"), Arrays.asList(canceledStatus, activeStatus));
+        } finally {
+            if (conferenceService.getConferenceByName("Test Conference") != null) {
+                conferenceService.deleteByName("Test Conference");
+            }
+        }
     }
 
     @Test
@@ -109,60 +120,109 @@ public class Tests {
         conferenceService.create(new Conference("Test Conference",
                 LocalDateTime.parse("2020-12-12 13:00", DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm"))));
 
-        driver.get("http://localhost:8080/conferences");
-        ConferencePage conferencePage = new ConferencePage(driver);
-        conferencePage.registerParticipant("Peter Nolan", "12-12-2000");
+        try {
+            driver.get("http://localhost:8080/conferences");
+            ConferencePage conferencePage = new ConferencePage(driver);
+            conferencePage.registerParticipant("Peter Nolan", "12-12-2000");
 
-        driver.get("http://localhost:8080/register");
-        RegisterPage registerPage = new RegisterPage(driver);
-        registerPage.conferences();
+            driver.get("http://localhost:8080/register");
+            RegisterPage registerPage = new RegisterPage(driver);
+            registerPage.conferences();
 
-        driver.get("http://localhost:8080/conferences");
-        conferencePage = new ConferencePage(driver);
-        conferencePage.showRegistered();
+            driver.get("http://localhost:8080/conferences");
+            conferencePage = new ConferencePage(driver);
+            conferencePage.showRegistered();
 
-        driver.get("http://localhost:8080/participants");
-        ParticipantsPage participantsPage = new ParticipantsPage(driver);
-        List name = participantsPage.participantData();
-        participantsPage.resetDeleteParticipant();
-        participantsPage.submitDeleteParticipant();
-        participantsPage.clickReturn();
+            driver.get("http://localhost:8080/participants");
+            ParticipantsPage participantsPage = new ParticipantsPage(driver);
+            List name = participantsPage.participantData();
+            participantsPage.submitDeleteParticipant();
+            participantsPage.clickReturn();
 
-        driver.get("http://localhost:8080/conferences");
-        conferencePage = new ConferencePage(driver);
-        conferencePage.deleteConference();
+            driver.get("http://localhost:8080/conferences");
+            conferencePage = new ConferencePage(driver);
+            conferencePage.deleteConference();
 
-        assertEquals(Arrays.asList("Peter Nolan", "2000-12-12"), name);
+            assertEquals(Arrays.asList("Peter Nolan", "2000-12-12"), name);
+        }finally {
+            if (conferenceService.getConferenceByName("Test Conference") != null) {
+                conferenceService.deleteByName("Test Conference");
+            }
+            if (conferenceParticipantService.findConferenceParticipantByFullName("Peter Nolan") != null) {
+                conferenceParticipantService.deleteConferenceParticipantByFullName("Peter Nolan");
+            }
+        }
     }
 
     @Test
     public void deleteConference() {
-        conferenceService.create(new Conference("Test Conference",
-                LocalDateTime.parse("2020-12-12 13:00", DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm"))));
-        driver.get("http://localhost:8080/conferences");
-        ConferencePage conferencePage = new ConferencePage(driver);
-        conferencePage.deleteConference();
+        try {
+            conferenceService.create(new Conference("Test Conference",
+                    LocalDateTime.parse("2020-12-12 13:00", DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm"))));
+
+            driver.get("http://localhost:8080/conferences");
+            ConferencePage conferencePage = new ConferencePage(driver);
+            conferencePage.deleteConference();
+        } finally {
+            if (conferenceService.getConferenceByName("Test Conference") != null) {
+                conferenceService.deleteByName("Test Conference");
+            }
+        }
     }
 
     @Test
     public void selectRoom() {
-        conferenceService.create(new Conference("Test Conference",
-                LocalDateTime.parse("2020-12-12 13:00", DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm"))));
-        conferenceRoomService.create(new ConferenceRoom("Test Room", "basement", 150));
+        try {
+            conferenceService.create(new Conference("Test Conference",
+                    LocalDateTime.parse("2020-12-12 13:00", DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm"))));
+            conferenceRoomService.create(new ConferenceRoom("Test Room", "basement", 150));
 
-        driver.get("http://localhost:8080/conferences");
-        ConferencePage conferencePage = new ConferencePage(driver);
-        conferencePage.selectRoom();
-        driver.get("http://localhost:8080/selectRoom");
-        SelectRoomPage selectRoomPage = new SelectRoomPage(driver);
-        selectRoomPage.submitSelectRoom();
-        driver.get("http://localhost:8080/conferences");
-        conferencePage = new ConferencePage(driver);
-        String actualRoom = conferencePage.checkSelectedRoom();
+            driver.get("http://localhost:8080/conferences");
+            ConferencePage conferencePage = new ConferencePage(driver);
+            conferencePage.selectRoom();
 
-        conferencePage.deleteConference();
-        conferenceRoomService.deleteConferenceRoomByRoomName("Test Room");
+            driver.get("http://localhost:8080/selectRoom");
+            SelectRoomPage selectRoomPage = new SelectRoomPage(driver);
+            selectRoomPage.submitSelectRoom();
 
-        assertEquals("Test Room", actualRoom);
+            driver.get("http://localhost:8080/conferences");
+            conferencePage = new ConferencePage(driver);
+            String actualRoom = conferencePage.checkSelectedRoom();
+
+            conferencePage.deleteConference();
+
+            assertEquals("Test Room", actualRoom);
+        } finally {
+            if (conferenceRoomService.getConferenceRoomByName("Test Room") != null) {
+                conferenceRoomService.deleteByName("Test Room");
+            }
+            if (conferenceService.getConferenceByName("Test Conference") != null) {
+                conferenceService.deleteByName("Test Conference");
+            }
+        }
+    }
+
+    @Test
+    public void addConferenceRoom() {
+        try {
+            driver.get("http://localhost:8080/conferenceRooms");
+            ConferenceRoomPage conferenceRoomPage = new ConferenceRoomPage(driver);
+            conferenceRoomPage.addConferenceRoom();
+
+            driver.get("http://localhost:8080/addConferenceRoom");
+            AddConferenceRoomPage addConferenceRoomPage = new AddConferenceRoomPage(driver);
+            addConferenceRoomPage.setData("Test Room", "Test Floor", "150");
+            addConferenceRoomPage.submit();
+
+            driver.get("http://localhost:8080/conferenceRooms");
+            conferenceRoomPage = new ConferenceRoomPage(driver);
+            List actualRoom = conferenceRoomPage.checkData();
+            assertEquals(Arrays.asList("Test Room", "Test Floor", "150"), actualRoom);
+        } finally {
+            if (conferenceRoomService.getConferenceRoomByName("Test Room") != null) {
+                conferenceRoomService.deleteByName("Test Room");
+            }
+        }
+
     }
 }
